Delete fruit manager and clean up interface injection pass
